requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==25.1.0
    # via nicegui
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via
    #   nicegui
    #   python-socketio
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncify==0.12.1
    # via maadebugger
attrs==25.4.0
    # via aiohttp
bidict==0.23.1
    # via python-socketio
black==25.11.0
    # via maagf2exilium (pyproject.toml)
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   nicegui
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   black
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
docutils==0.22.3
    # via nicegui
fastapi==0.123.5
    # via nicegui
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
funkify==0.7.1
    # via asyncify
h11==0.16.0
    # via
    #   httpcore
    #   nicegui
    #   uvicorn
    #   wsproto
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via nicegui
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ifaddr==0.2.0
    # via nicegui
itsdangerous==2.2.0
    # via nicegui
jinja2==3.1.6
    # via nicegui
maaagentbinary==1.0.1
    # via maafw
maadebugger==1.14.1
    # via maagf2exilium (pyproject.toml)
maafw==5.1.3
    # via
    #   maagf2exilium (pyproject.toml)
    #   maadebugger
markdown2==2.5.4
    # via nicegui
markupsafe==3.0.3
    # via jinja2
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via black
nicegui==2.24.2
    # via maadebugger
numpy==2.3.5
    # via maafw
orjson==3.11.4
    # via nicegui
packaging==25.0
    # via
    #   black
    #   maadebugger
pathspec==0.12.1
    # via black
pillow==12.0.0
    # via maadebugger
platformdirs==4.5.0
    # via black
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pscript==0.7.7
    # via vbuild
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via nicegui
python-dotenv==1.2.1
    # via uvicorn
python-engineio==4.12.3
    # via
    #   nicegui
    #   python-socketio
python-multipart==0.0.20
    # via nicegui
python-socketio==5.15.0
    # via nicegui
pytokens==0.3.0
    # via black
pyyaml==6.0.3
    # via uvicorn
requests==2.32.5
    # via maagf2exilium (pyproject.toml)
simple-websocket==1.1.0
    # via python-engineio
starlette==0.50.0
    # via
    #   fastapi
    #   nicegui
strenum==0.4.15
    # via maafw
tqdm==4.67.1
    # via maagf2exilium (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   nicegui
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.38.0
    # via nicegui
vbuild==0.8.2
    # via nicegui
watchfiles==1.1.1
    # via
    #   nicegui
    #   uvicorn
websockets==15.0.1
    # via uvicorn
wsproto==1.3.2
    # via simple-websocket
yarl==1.22.0
    # via aiohttp
